pyproject.toml
[project]
name = "ExpAgent"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "kaggle>=1.8.4",
    "google-cloud-aiplatform>=1.74.0",
    "nbformat>=5.10.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.1",
    "rootutils>=1.0.7",
    "tyro>=1.0.6",
    "lightgbm>=4.6.0",
    "scikit-learn>=1.8.0",
    "pandas>=2.2.0,<3.0.0",
    "stickytape>=0.2.1",
    "mlflow>=2.0.0",
]

[dependency-groups]
dev = ["pytest>=9.0.2", "ruff>=0.14.13", "ty>=0.0.12"]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [".claude/"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
    "B905", # zip without explicit strict parameter
    "E402",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
known-local-folder = ["settings", "train"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
